Pretty good framework laid out
I made a grid system, along an UI system and a system to place blocks

The grid system is currently 5x5 (At first I said 10x10 but I didn't know how big that was so I dropped it down to 5x5, you can change the parameters on the globals variables I put on top). You can navigate through the layers by using W/S to move up and down the vertical.

The UI system is just ControlP5 with two toggle buttons to change from brush to eraser along with a color wheel (that is kinda buggy because of the library idk?)

Placing blocks is pretty simple. I did a arrow key system at first but I realized that was too annoying to navigate so I swapped to this layer by layer editing system, where you can add blocks to the map on the top-left can switch between layers with W and S