spheres!
I added spheres, however there are a few bugs I didn't fix, namely:
a. You can draw cubes and spheres on the same tile which is not necessarily bad but idk if we want it
b. I couldn't figure out how to get spheres and cube indicators to show up together on the top down view